requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
aiosqlite>=0.20.0
qrcode[pil]>=8.0
Jinja2>=3.1.5
python-dotenv>=1.0.0
itsdangerous>=2.1.0
